Implement Book Specification based on the Physical Product class.
[BusinessRuleEngine]

Implement the Book specififcation. The book is a phyiscal product so inherited from
the PhyiscalProduct.
Create unit test for the Book. All the information is contained within the PackingSlip.
If the payment is for a physical product, generate a packing slip for shipping.
If the payment is for a book, create a duplicate packing slip for the royalty department.

[Tested by]
Jody cole

[Code review]
Jody cole