Membership class work. Start thinking about finishing up tidy up code and more comments.
[BusinessRuleEngine]

Work on the Membership class. This is bit more tricky then others before.
We have to think about incorporating the owner into the membership.
Create Owner class with email, name. Email is taken as the unique for the owner not name.
Owner defines isMember which is true if this owner is a member and isUpgraded - true
if owner is already a member and upgraded.
Write unit test for Membership to ensure spec is fullfiled and wont be broken in future
maitenance.

[Tested by]
Jody cole

[Code review]
Jody cole